Make sure mineral gatherers are gethering minerals each frame
There was a bug with Zerg where the hatching drones
wouldn't gather minerals. It appears to be because
there is a small amount of time where the newly
hatched drone can't take a command. This has been
corrected by a general solution in which each frame the
ResourceManager makes sure all it's managed workers
are gathering resources.