Removed the boundary classes.
The boundary classes were a hinderence due to the large number
of constants contained in {Upgrade,Unit,Tech,...}Type. To prevent
this from getting out of control I've removed the boundary classes
in general. This will also prevent issues due to caching.

It may be valuable to revive the Unit boundary class in the future.